requirements.txt
python-telegram-bot[job-queue]==21.6
aiosqlite==0.20.0
feedparser==6.0.11
requests==2.33.1
deep-translator==1.11.4
langdetect==1.0.9
python-dotenv==1.0.1
urllib3==2.2.3
